backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
pandas
openpyxl
langchain-core
langchain-openai
pydantic
python-multipart
